+++ src/style/padding.rs
@@ -1,3 +1,5 @@
+use crate::utils::floats::Float;
+
 /// Represents the padding values for a UI element.
 ///
 /// Padding can be set individually for each side (left, right, top, bottom).
@@ -7,15 +9,15 @@
 /// For more info on Styling, see [`kaolin::style`]
 #[derive(Default, Clone, Copy)]
 pub struct Padding {
-    pub left: f64,
-    pub right: f64,
-    pub top: f64,
-    pub bottom: f64,
+    pub left: Float,
+    pub right: Float,
+    pub top: Float,
+    pub bottom: Float,
 }
 
 impl Padding {
     /// Creates a new `Padding` instance with the specified values.
-    pub fn new(left: f64, right: f64, top: f64, bottom: f64) -> Self {
+    pub fn new(left: Float, right: Float, top: Float, bottom: Float) -> Self {
         Padding {
             left,
             right,
@@ -25,7 +27,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` instance with the same value for all sides.
-    pub fn all(value: f64) -> Self {
+    pub fn all(value: Float) -> Self {
         Padding {
             left: value,
             right: value,
@@ -35,7 +37,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` instance with the same value for the horizontal sides (left and right).
-    pub fn horizontal(value: f64) -> Self {
+    pub fn horizontal(value: Float) -> Self {
         Padding {
             left: value,
             right: value,
@@ -45,7 +47,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` instance with the same value for the vertical sides (top and bottom).
-    pub fn vertical(value: f64) -> Self {
+    pub fn vertical(value: Float) -> Self {
         Padding {
             left: 0.0,
             right: 0.0,
@@ -55,7 +57,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` with the same value for the horizontal sides (left and right) and the vertical sides (top and bottom).
-    pub fn hor_ver(hor: f64, ver: f64) -> Self {
+    pub fn hor_ver(hor: Float, ver: Float) -> Self {
         Padding {
             left: hor,
             right: hor,
@@ -65,7 +67,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` with only the left value set, all others sides set to 0.
-    pub fn left(value: f64) -> Self {
+    pub fn left(value: Float) -> Self {
         Padding {
             left: value,
             right: 0.0,
@@ -75,7 +77,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` with only the right value set, all others sides set to 0.
-    pub fn right(value: f64) -> Self {
+    pub fn right(value: Float) -> Self {
         Padding {
             left: 0.0,
             right: value,
@@ -85,7 +87,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` with only the top value set, all others sides set to 0.
-    pub fn top(value: f64) -> Self {
+    pub fn top(value: Float) -> Self {
         Padding {
             left: 0.0,
             right: 0.0,
@@ -95,7 +97,7 @@ impl Padding {
     }
 
     /// Creates a new `Padding` with only the bottom value set, all others sides set to 0.
-    pub fn bottom(value: f64) -> Self {
+    pub fn bottom(value: Float) -> Self {
         Padding {
             left: 0.0,
             right: 0.0,
@@ -105,12 +107,14 @@ impl Padding {
     }
 
     /// Returns the total horizontal padding (left + right). Used internally to calculate layout positions.
-    pub fn x(&self) -> f64 {
+    #[inline]
+    pub fn x(&self) -> Float {
         self.left + self.right
     }
 
     /// Returns the total vertical padding (top + bottom). Used internally to calculate layout positions.
-    pub fn y(&self) -> f64 {
+    #[inline]
+    pub fn y(&self) -> Float {
         self.top + self.bottom
     }
 }
